+++ src/pages/Executors.vue
@@ -1,6 +1,6 @@
 <template>
   <q-page class="q-pa-md">
-    <q-breadcrumbs active-color="secondary" color="light">
+    <q-breadcrumbs active-color="secondary" color="light" class="q-mb-sm">
       <q-breadcrumbs-el label="Home" to="/"/>
       <q-btn
         dense round icon="refresh" class="q-ml-lg" size="sm"